+++ src/main/java/org/idnt/udemy/springboot/app/example/controller/dto/TransactionDTO.java
@@ -1,0 +1,17 @@
+package org.idnt.udemy.springboot.app.example.controller.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class TransactionDTO {
+    private Long idBank;
+    private Long idAccountOrigin;
+    private Long idAccountTarget;
+    private BigDecimal quantity;
+}