+++ src/main/java/org/idnt/udemy/springboot/app/example/service/impl/AccountServiceImpl.java
@@ -8,8 +8,10 @@ import org.idnt.udemy.springboot.app.example.repository.AccountRepository;
 import org.idnt.udemy.springboot.app.example.repository.BankRepository;
 import org.idnt.udemy.springboot.app.example.service.AccountService;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
+import java.util.List;
 
 @Data
 @AllArgsConstructor
@@ -19,23 +21,27 @@ public class AccountServiceImpl implements AccountService {
     private BankRepository bankRepository;
 
     @Override
+    @Transactional(readOnly = true)
     public Account findById(final Long idAccount) {
         return this.accountRepository.findById(idAccount).orElseThrow();
     }
 
     @Override
+    @Transactional(readOnly = true)
     public int checkTotalTransactions(final Long idBank) {
         Bank bank = this.bankRepository.findById(idBank).orElseThrow();
         return bank.getTotalTransactions();
     }
 
     @Override
+    @Transactional(readOnly = true)
     public BigDecimal checkBalance(final Long idAccount) {
         Account account = this.accountRepository.findById(idAccount).orElseThrow();
         return account.getBalance();
     }
 
     @Override
+    @Transactional
     public void transfer(final Long idBank, final Long idAccountOrigin, final Long idAccountTarget, final BigDecimal quantity) {
         Account accountOrigin = this.accountRepository.findById(idAccountOrigin).orElseThrow();
         accountOrigin.debit(quantity);
@@ -49,5 +55,17 @@ public class AccountServiceImpl implements AccountService {
         int totalTransactions = bank.getTotalTransactions();
         bank.setTotalTransactions(++totalTransactions);
         this.bankRepository.save(bank);
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public List<Account> findAll() {
+        return this.accountRepository.findAll();
+    }
+
+    @Transactional
+    @Override
+    public Account save(Account account) {
+        return this.accountRepository.save(account);
     }
 }