+++ src/main/resources/import.sql
@@ -1,0 +1,2 @@
+INSERT INTO accounts (person_name, balance) VALUES ('Andrés', 1000), ('Jhon', 2000);
+INSERT INTO banks (name, total_transactions) VALUES ('El banco financiero', 0);