+++ src/test/java/org/idnt/udemy/springboot/app/example/test/SpringbootTestApplicationTests.java
@@ -14,6 +14,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
 import java.math.BigDecimal;
+import java.util.Arrays;
+import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -144,6 +146,46 @@ class SpringbootTestApplicationTests {
 		verify(this.bankRepository, never()).findById(anyLong());
 		verify(this.bankRepository, never()).save(any(Bank.class));
 		verify(this.bankRepository, never()).findAll();
+	}
 
+	@Test
+	@DisplayName("Test to check that the account service retrieves the list of accounts.")
+	void testFindAll() {
+		//Given
+		final List<Account> ACCOUNTS_EXPECTED = Arrays.asList(DATA.getAccount001().get(), DATA.getAccount002().get());
+		when(this.accountRepository.findAll()).thenReturn(ACCOUNTS_EXPECTED);
+
+		//When
+		final List<Account> RESULT = this.accountService.findAll();
+
+		//Then
+		assertNotNull(RESULT, () -> "The result mustn't be null");
+		assertFalse(RESULT.isEmpty(), () -> "The result mustn't be empty");
+		assertEquals(2, RESULT.size());
+		assertTrue(RESULT.contains(DATA.getAccount001().get()), () -> String.format("Te result must contain %s", DATA.getAccount001()));
+		verify(this.accountRepository).findAll();
+	}
+
+	@Test
+	@DisplayName("Test to check that the account service holds an account for the account.")
+	void testSave() {
+		//Given
+		final Long ID = 3L;
+		final Account NEW_ACCOUNT = new Account(null, "Patricia", new BigDecimal("3000"));
+		when(this.accountRepository.save(any())).then(invocation -> {
+			Account accountArg = invocation.getArgument(0);
+			accountArg.setId(ID);
+			return accountArg;
+		});
+
+		//When
+		final Account RESULT = this.accountService.save(NEW_ACCOUNT);
+
+		//Then
+		assertNotNull(RESULT, () -> "The result mustn't be null");
+		assertEquals("Patricia", RESULT.getPersonName());
+		assertEquals(ID, RESULT.getId());
+		assertEquals("3000", RESULT.getBalance().toPlainString());
+		verify(this.accountRepository).save(any());
 	}
 }